Move javascript into external file, and avoid meta-refresh.
We've moved our core javascript into `/purple.js`, and trigger
it to be loaded via AJAX every 30 seconds.  This replaces the use
of META-REFRESH to update the page every minute, which is neater.

NOTE: This causes a bug as events are constantly appended to
our tables, we'll fix that in a separate update.